server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
requests==2.30.0
python-dotenv==1.0.0
